+++ app/styles/app.scss
@@ -40,4 +40,3 @@ th.numeric, td.numeric{
 
   a { color: white !important; }
 }
-
