--- app/styles/test.scss
@@ -1,7 +1,0 @@
-#scratch {
-  color: black !important;
-
-  table { color: black !important; }
-
-  a { color: black !important; }
-}
