--- app/styles/vendor.scss
@@ -1,1 +1,0 @@
-@import "bower_components/bootstrap-sass/assets/stylesheets/bootstrap";
