+++ app/templates/application.hbs
@@ -1,0 +1,4 @@
+<div class="container">
+  <h3 class="text-muted">Ember Performance</h3>
+  {{outlet}}
+</div>
