+++ benchmarks/baseline-handlebars-list/index.js
