+++ benchmarks/ember-get/primed/index.js
