+++ benchmarks/object-create/baseline/index.js
