+++ benchmarks/object-create/component/index.js
