+++ benchmarks/object-create/index.js
