+++ benchmarks/object-create/without-non-enumerable-safety-same-class/index.js
