+++ benchmarks/object-create/without-non-enumerable-safety/index.js
