+++ benchmarks/render-bind-attr/index.js
