+++ benchmarks/render-simple-ember-list/index.js
