+++ bower.json
@@ -5,20 +5,23 @@
     "Robin Ward <robin.ward@gmail.com>"
   ],
   "license": "MIT",
-  "ignore": [
-    "**/.*",
-    "node_modules",
-    "bower_components",
-    "test",
-    "tests"
-  ],
   "dependencies": {
-    "bootstrap-sass": "~3.3.5",
-    "jquery": "~2.1.1",
+    "ascii-table": "~0.0.4",
     "benchmark": "git@github.com:bestiejs/benchmark.js.git",
+    "ember": "1.13.10",
+    "ember-cli-shims": "ember-cli/ember-cli-shims#0.0.3",
+    "ember-cli-test-loader": "ember-cli-test-loader#0.1.3",
+    "ember-data": "1.13.8",
+    "ember-load-initializers": "ember-cli/ember-load-initializers#0.1.5",
+    "ember-qunit": "0.4.9",
+    "ember-qunit-notifications": "0.0.7",
+    "ember-resolver": "~0.1.18",
     "headjs": "~1.0.3",
     "rsvp": "3.0.13",
-    "ascii-table": "~0.0.4",
-    "numeraljs": "~1.5.3"
+    "jquery": "~2.1.1",
+    "loader.js": "ember-cli/loader.js#3.2.1",
+    "numeraljs": "~1.5.3",
+    "qunit": "~1.18.0",
+    "bootstrap": "~3.3.2"
   }
 }
