+++ package.json
@@ -1,24 +1,47 @@
 {
-  "name": "ember-performance-jjabrams",
-  "version": "0.9.2",
+  "name": "ember-performance",
+  "version": "0.10.0",
+  "description": "Benchmark Ember.js Performance",
+  "license": "MIT",
+  "private": true,
+  "directories": {
+    "doc": "doc",
+    "test": "tests"
+  },
+  "scripts": {
+    "build": "ember build",
+    "start": "ember server",
+    "test": "ember test"
+  },
+  "repository": "",
+  "engines": {
+    "node": ">= 0.10.0"
+  },
   "author": "Robin Ward",
   "license": "MIT",
-  "scripts": {
-    "start": "ember server",
-    "postinstall": "bower install"
-  },
   "devDependencies": {
+    "broccoli-asset-rev": "^2.1.2",
     "broccoli-bower": "^0.2.1",
-    "broccoli-clean-css": "^0.3.0",
     "broccoli-concat": "0.0.12",
-    "broccoli-env": "0.0.1",
     "broccoli-filter": "^0.1.10",
-    "broccoli-merge-trees": "^0.2.1",
-    "broccoli-sane-watcher": "^0.1.0",
-    "broccoli-sass": "^0.6.2",
-    "broccoli-static-compiler": "^0.2.1",
-    "broccoli-uglify-js": "^0.1.3",
-    "ember-cli": "^1.13.8",
-    "broccoli-stew": "^0.1.2"
+    "broccoli-funnel": "^0.2.8",
+    "broccoli-merge-trees": "^0.2.3",
+    "broccoli-stew": "^0.1.2",
+    "ember-bootstrap": "0.4.0",
+    "ember-cli": "1.13.8",
+    "ember-cli-app-version": "0.5.0",
+    "ember-cli-babel": "^5.1.3",
+    "ember-cli-dependency-checker": "^1.0.1",
+    "ember-cli-htmlbars": "0.7.9",
+    "ember-cli-htmlbars-inline-precompile": "^0.2.0",
+    "ember-cli-ic-ajax": "0.2.1",
+    "ember-cli-inject-live-reload": "^1.3.1",
+    "ember-cli-qunit": "^1.0.0",
+    "ember-cli-release": "0.2.3",
+    "ember-cli-sass": "5.0.0",
+    "ember-cli-sri": "^1.0.3",
+    "ember-cli-uglify": "^1.2.0",
+    "ember-disable-proxy-controllers": "^1.0.0",
+    "ember-export-application-global": "^1.0.3"
   }
 }
