+++ test-client/index-template.html
@@ -2,8 +2,7 @@
   <head>
     <title>Ember Performance - Testing</title>
     <link rel="stylesheet" href="/assets/vendor.css">
-    <link rel="stylesheet" href="/assets/app.css">
-    <link rel="stylesheet" href="/assets/test.css">
+    <link rel="stylesheet" href="/assets/ember-performance.css">
     <meta name="viewport" content="width=device-width, initial-scale=1">
   </head>
